requirements.txt
discord.py==1.7.3
youtube_dl==2021.12.17
ffmpeg==1.4
PyNaCl==1.5.0
